requirements.txt
fastapi
uvicorn
ibis-framework[duckdb]
docopt